Refactor code. Unusable code was deleted
