--- src/main/java/page/TestData.java
@@ -1,7 +1,0 @@
-package page;
-
-public class TestData {
-
-
-
-}
