+++ src/test/java/lessons/AddLesson_VerifyAddLesson_CorrectData_MentorRole.java
@@ -3,7 +3,7 @@ package lessons;
 import constants.Endpoints;
 import lessons.data.AddLessonData;
 import org.testng.annotations.BeforeClass;
-import unauthorizedPages.Auth;
+import page.unauthorizedUserPages.AuthPage;
 import page.lessons.LessonsPage;
 import util.Role;
 import java.io.IOException;
@@ -17,7 +17,7 @@ public class AddLesson_VerifyAddLesson_CorrectData_MentorRole extends AddLesson_
     @Override
     public void precondition() throws IOException {
         driver.get(Endpoints.BASE_URL);
-        lessonsPage = Auth.init(driver)
+        lessonsPage = AuthPage.init(driver)
                 .logInAs(Role.MENTOR, LessonsPage.class)
                 .isAtPage(waitTime);
     }
